+++ src/transformers/byteSequenceShaper.ts
