+++ src/transformers/decompressionShaper.ts
@@ -2,7 +2,7 @@
 
 import arithmetic = require('./arithmetic');
 import arraybuffers = require('../arraybuffers/arraybuffers');
-import decompression = require('../fancy-transformers/decompressionShaper');
+import decompression = require('./decompressionShaper');
 import logging = require('../logging/logging');
 
 const log :logging.Log = new logging.Log('compressionShaper');
